refactor example script runner to helper
with this it will be possible to resue the runner and split up
examples to smaller testsuites if needed
