+++ tests/functional/examples/test_examples.py
@@ -1,20 +1,10 @@
 import os
-import re
-import runpy
-import sys
-from traceback import format_exc
 
 import pytest
 
-from testplan.common.utils.path import change_directory
-
 import platform
 
-SUCCES_EXIT_CODES = (
-    0,
-    False,
-    None,
-)  # these considerd as 0 return value and means success
+from example_runner import run_example_in_process
 
 _FILE_DIR = os.path.dirname(__file__)
 
@@ -90,58 +80,4 @@ def test_example(root, filename):
     elif any([file_path.endswith(skip_name) for skip_name in SKIP]):
         pytest.skip()
 
-    try:
-
-        # set up like an external invocation.
-        # add current dir to fron of sys path
-        # no arguments pass in command line
-
-        sys.path.insert(0, root)
-        argv = sys.argv
-        sys.argv = [sys.argv[0]]
-
-        with change_directory(root), open(filename) as file_obj:
-            file_obj.readline()
-            second_line = file_obj.readline()
-
-            runpy.run_path(os.path.join(root, filename), run_name="__main__")
-
-    except SystemExit as e:
-        if e.code not in SUCCES_EXIT_CODES:
-            assert (
-                "# This plan contains tests that demonstrate failures as well."
-            ) == second_line.strip(), (
-                "Expected '{}' example to pass, it failed".format(file_path)
-            )
-    except Exception as e:
-        for exception in KNOWN_EXCEPTIONS:
-            if re.search(exception, "{}: {}".format(type(e).__name__, str(e))):
-                pytest.xfail()
-        pytest.fail(format_exc())
-    finally:
-        # clean up after execution
-        # remove the modules imported from the directory of testplan_path
-        # remove the added toot from sys.path
-        # reset the args as it was before execution
-
-        for m in get_local_modules(root):
-            del sys.modules[m]
-
-        sys.path = sys.path[1:]
-        sys.argv = argv
-
-
-def get_local_modules(path_prefix):
-    """
-    return module names loaded from a path with path_prefix
-
-    :param path_prefix:
-    :return:
-    """
-    return set(
-        n
-        for n, m in sys.modules.items()
-        if hasattr(m, "__file__")
-        and m.__file__
-        and m.__file__.startswith(path_prefix)
-    )
+    run_example_in_process(filename, root, KNOWN_EXCEPTIONS)
